chore: add modules/category dir and move to it the related files
move related types in additional.ts file to each category file
