+++ frontend/src/modules/category/components/card/index.ts
@@ -1,0 +1,2 @@
+export * from './ProductCard.category'
+export * from './ProductDiscountInfo.category'
