+++ frontend/src/modules/category/components/filters/index.ts
@@ -1,0 +1,4 @@
+export * from './SearchFilter.category'
+export * from './BrandFilter.category'
+export * from './SortFilter.category'
+export * from './DiscountFilter.category'
